test(sdk): add a standalone conformance server and runner
